Split out reading system probe config to account for unsupported builds
