+++ cmd/agent/common/system_probe_config_helper_unsupported.go
@@ -1,0 +1,8 @@
+// +build !process windows
+
+package common
+
+// SetupSystemProbeConfig returns nil on unsupported builds
+func SetupSystemProbeConfig(sysProbeConfFilePath string) error {
+	return nil
+}
